+++ StreamChatSwiftUITests/Tests/ChatChannel/MessageContainerView_Tests.swift
@@ -33,6 +33,25 @@ class MessageContainerView_Tests: StreamChatTestCase {
         // Then
         assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
     }
+    
+    func test_messageContainerEdited_snapshot() {
+        // Given
+        streamChat = StreamChat(chatClient: chatClient)
+        let message = ChatMessage.mock(
+            id: .unique,
+            cid: .unique,
+            text: "Message sent by current user",
+            author: .mock(id: Self.currentUserId, name: "Martin"),
+            isSentByCurrentUser: true,
+            textUpdatedAt: Date()
+        )
+
+        // When
+        let view = testMessageViewContainer(message: message)
+
+        // Then
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+    }
 
     func test_messageContainerViewSentOtherUser_snapshot() {
         // Given